Fix several long long warnings
git-svn-id: https://svn.dealii.org/branches/branch_bigger_global_dof_indices_4@29214 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/include/deal.II/base/partitioner.h
@@ -448,7 +448,7 @@ namespace Utilities
         return static_cast<unsigned int>(global_index - local_range_data.first);
       else if (is_ghost_entry (global_index))
         return (local_size() +
-                ghost_indices_data.index_within_set (global_index));
+                static_cast<unsigned int>(ghost_indices_data.index_within_set (global_index)));
       else
         // should only end up here in
         // optimized mode, when we use this
